+++ features/conversation/conversation-message/conversation-message-reactions/conversation-message-reaction-drawer/conversation-message-reaction-drawer.service.ts
@@ -1,17 +1,20 @@
+import { MessageId } from "@xmtp/react-native-sdk";
 import {
   resetMessageReactionsStore,
   useMessageReactionsStore,
 } from "./conversation-message-reaction-drawer.store";
-import { RolledUpReactions } from "../conversation-message-reactions.types";
+import { createBottomSheetModalRef } from "@/design-system/BottomSheet/BottomSheet.utils";
 
-export function openMessageReactionsDrawer(reactions: RolledUpReactions) {
-  useMessageReactionsStore.getState().setRolledUpReactions(reactions);
+export const conversationMessageDrawerBottomSheetRef =
+  createBottomSheetModalRef();
+
+export function openMessageReactionsDrawer(args: { messageId: MessageId }) {
+  const { messageId } = args;
+  const store = useMessageReactionsStore.getState();
+  store.actions.setMessageId(messageId);
 }
 
 export function closeMessageReactionsDrawer() {
+  conversationMessageDrawerBottomSheetRef.current?.close();
   resetMessageReactionsStore();
-}
-
-export function useMessageReactionsRolledUpReactions() {
-  return useMessageReactionsStore((state) => state.rolledUpReactions);
 }
