+++ features/conversation/conversation-message/conversation-message-reactions/conversation-message-reaction-drawer/conversation-message-reaction-drawer.store.ts
@@ -1,33 +1,24 @@
-import { create } from "zustand";
-import { RolledUpReactions } from "../conversation-message-reactions.types";
 import { MessageId } from "@xmtp/react-native-sdk";
+import { create } from "zustand";
 
 export type IMessageReactionsStore = {
-  rolledUpReactions: RolledUpReactions;
-  setRolledUpReactions: (reactions: RolledUpReactions) => void;
+  messageId: MessageId | null;
+  actions: {
+    setMessageId: (messageId: MessageId | null) => void;
+  };
 };
 
 export const useMessageReactionsStore = create<IMessageReactionsStore>(
   (set) => ({
-    rolledUpReactions: {
-      preview: [],
-      detailed: [],
-      totalCount: 0,
-      userReacted: false,
-      messageId: "" as MessageId,
+    messageId: null,
+    actions: {
+      setMessageId: (messageId) => set({ messageId }),
     },
-    setRolledUpReactions: (reactions) => set({ rolledUpReactions: reactions }),
   })
 );
 
 export function resetMessageReactionsStore() {
   useMessageReactionsStore.setState({
-    rolledUpReactions: {
-      preview: [],
-      detailed: [],
-      totalCount: 0,
-      userReacted: false,
-      messageId: "" as MessageId,
-    },
+    messageId: null,
   });
 }
