+++ features/conversation/conversation-message/conversation-message-reactions/conversation-message-reactions.tsx
@@ -1,26 +1,13 @@
 import { useSelect } from "@/data/store/storeHelpers";
 import { useMessageContextStoreContext } from "@/features/conversation/conversation-message/conversation-message.store-context";
-import { useConversationMessageReactions } from "@/features/conversation/conversation-message/conversation-message.utils";
-import { isCurrentAccountInboxId } from "@/hooks/use-current-account-inbox-id";
-import { getInboxProfileSocialsQueryConfig } from "@/queries/useInboxProfileSocialsQuery";
 import { AnimatedHStack, HStack } from "@design-system/HStack";
 import { Text } from "@design-system/Text";
 import { VStack } from "@design-system/VStack";
-import { useQueries } from "@tanstack/react-query";
 import { ThemedStyle, useAppTheme } from "@theme/useAppTheme";
-import {
-  getPreferredInboxAddress,
-  getPreferredInboxAvatar,
-  getPreferredInboxName,
-} from "@utils/profile";
-import { memo, useCallback, useMemo } from "react";
+import { memo, useCallback } from "react";
 import { TextStyle, TouchableHighlight, ViewStyle } from "react-native";
 import { openMessageReactionsDrawer } from "./conversation-message-reaction-drawer/conversation-message-reaction-drawer.service";
-import {
-  RolledUpReactions,
-  SortedReaction,
-} from "./conversation-message-reactions.types";
-import { debugBorder } from "@/utils/debug-style";
+import { useConversationMessageReactionsRolledUp } from "./use-conversation-message-reactions-rolled-up";
 
 const MAX_REACTION_EMOJIS_SHOWN = 3;
 
@@ -32,14 +19,15 @@ export const ConversationMessageReactions = memo(
       useSelect(["fromMe", "messageId"])
     );
 
-    const rolledUpReactions = useMessageReactionsRolledUp();
+    const rolledUpReactions = useConversationMessageReactionsRolledUp({
+      messageId: messageId,
+    });
 
     const handlePressContainer = useCallback(() => {
       openMessageReactionsDrawer({
-        ...rolledUpReactions,
         messageId,
       });
-    }, [rolledUpReactions, messageId]);
+    }, [messageId]);
 
     if (rolledUpReactions.totalCount === 0) {
       return null;
@@ -87,97 +75,6 @@ export const ConversationMessageReactions = memo(
     );
   }
 );
-
-function useMessageReactionsRolledUp() {
-  const { messageId } = useMessageContextStoreContext(useSelect(["messageId"]));
-  const { bySender: reactionsBySender } =
-    useConversationMessageReactions(messageId);
-  const inboxIds = Array.from(
-    new Set(
-      Object.entries(reactionsBySender ?? {}).map(
-        ([senderInboxId]) => senderInboxId
-      )
-    )
-  );
-
-  const inboxProfileSocialsQueries = useQueries({
-    queries: inboxIds.map((inboxId) =>
-      getInboxProfileSocialsQueryConfig({ inboxId })
-    ),
-  });
-
-  const membersSocials = inboxProfileSocialsQueries.map(
-    ({ data: socials }, index) => {
-      return {
-        inboxId: inboxIds[index],
-        address: getPreferredInboxAddress(socials),
-        uri: getPreferredInboxAvatar(socials),
-        name: getPreferredInboxName(socials),
-      };
-    }
-  );
-
-  return useMemo((): RolledUpReactions => {
-    const detailed: SortedReaction[] = [];
-    let totalCount = 0;
-    let userReacted = false;
-
-    // Flatten reactions and track sender addresses
-    const flatReactions = Object.entries(reactionsBySender ?? {}).flatMap(
-      ([senderInboxId, senderReactions]) =>
-        senderReactions.map((reaction) => ({ senderInboxId, ...reaction }))
-    );
-    totalCount = flatReactions.length;
-
-    // Track reaction counts for preview
-    const previewCounts = new Map<string, number>();
-
-    flatReactions.forEach((reaction) => {
-      const isOwnReaction = isCurrentAccountInboxId(reaction.senderInboxId);
-
-      if (isOwnReaction) {
-        userReacted = true;
-      }
-
-      // Count reactions for the preview
-      previewCounts.set(
-        reaction.content,
-        (previewCounts.get(reaction.content) || 0) + 1
-      );
-
-      const socialDetails = membersSocials.find(
-        (social) => social.inboxId === reaction.senderInboxId
-      );
-
-      detailed.push({
-        content: reaction.content,
-        isOwnReaction,
-        reactor: {
-          address: reaction.senderInboxId,
-          userName: socialDetails?.name,
-          avatar: socialDetails?.uri,
-        },
-      });
-    });
-
-    // Sort detailed array to place all OWN reactions at the beginning
-    detailed.sort((a, b) => Number(b.isOwnReaction) - Number(a.isOwnReaction));
-
-    // Convert previewCounts map to array with content and count
-    const preview = Array.from(previewCounts, ([content, count]) => ({
-      content,
-      count,
-    }));
-
-    return {
-      totalCount,
-      userReacted,
-      preview,
-      detailed,
-      messageId,
-    };
-  }, [reactionsBySender, membersSocials, messageId]);
-}
 
 const $reactionButton: ThemedStyle<ViewStyle> = ({
   colors,
